+++ crates/goose/src/agents/mod.rs
@@ -8,6 +8,7 @@ pub mod prompt_manager;
 mod reply_parts;
 mod router_tool_selector;
 mod router_tools;
+mod schedule_tool;
 
 mod tool_execution;
 mod tool_router_index_manager;
